fix: moved logic to hook
